requirements.txt
python-dotenv==1.0.1
azure-core
azure-storage-blob==12.20.0